pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "larops"
version = "0.1.0"
description = "Laravel-first server operations CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.10,<3.0",
  "pyyaml>=6.0,<7.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.9,<1.0",
]

[project.scripts]
larops = "larops.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

